Add files via upload
NOTE: these files were made in a package-based IDE so when trying to use these files make sure that the package names are changed.  

The FileScanner Class -> Used for parsing through the files to ensure that each file is read in its entirety, and allocated the right resources to the right places. 
The Main Class -> Reads in the .RUN files and parses the .GRP and .SEC files at the same time (Buggy)
The TestFileScanner class -> tests the methods of the FileScanner class